requirements.txt
python-docx>=1.1.0
tqdm>=4.66.0
numpy>=1.26.0,<2.0
pydantic>=2.8.0
python-dotenv>=1.0.1

langchain==1.2.10
langchain-core==1.2.18
langchain-community==0.4.1
langchain-text-splitters==1.1.1
langchain-ollama==1.0.1

faiss-cpu>=1.8.0
rank-bm25>=0.2.2
transformers>=4.44.0,<5.0
sentence-transformers>=3.0.0,<4.0
torch>=2.2.0,<2.4
